moved timeline graph to its own file
